refactor: IssuesProvider 생성
