+++ src/main.tsx
@@ -2,10 +2,13 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 
 import App from './App.tsx';
+import { IssuesProvider } from './context/IssuesProvider.tsx';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
-    <App />
+    <IssuesProvider>
+      <App />
+    </IssuesProvider>
   </React.StrictMode>,
 );
